requirements.txt
requests
jmespath
python-dotenv

pandas
plotly
streamlit-aggrid